phase1 done
