+++ lib/view/widgets/CustomError.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+void showGlobalSnackBar(BuildContext context, String message) {
+  final snackBar = SnackBar(
+    content: Text(message),
+    duration: Duration(seconds: 1),
+    behavior: SnackBarBehavior.floating,
+    shape: RoundedRectangleBorder(
+      borderRadius: BorderRadius.circular(15.0),
+    ),
+    elevation: 4.0,
+    margin: EdgeInsets.only(bottom: 90.0, left: 20.0, right: 20.0),
+  );
+
+  ScaffoldMessenger.of(context).hideCurrentSnackBar();
+  ScaffoldMessenger.of(context).showSnackBar(snackBar);
+}
